pyproject.toml
[project]
name = "firt-agent-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
